mcp-vector/pyproject.toml
[tool.poetry]
name = "mcp-vector"
version = "1.0.0"
description = "MCP Vector Server - Exposes ChromaDB operations as MCP tools for RAG workflows"
authors = ["Nicolas Lallier"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
httpx = "^0.28.0"
chromadb = "^1.0.0"
mcp = "^1.0.0"
sse-starlette = "^2.2.0"
structlog = "^24.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-instrumentation-fastapi = "^0.50b0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
